separate digitalocean and docker
