+++ Pulumi.docker-prod.yaml
@@ -1,0 +1,3 @@
+config:
+  o-p.n:traefik_admin_creds:
+    secure: AAABAGa+R93S6jPcxq/w2YHa6LJPcfzHvYSR+5YXe91YC8owMh5gCKamrC1u6e+ZsQnrn51pNBtgs5yNM8F+Aw3uDteVcEO97Z2A7aPSygfPhkgyIDlLsnXYUHYyQGnkwn8=
